+++ src/Component/TaggedServices/TaggedServicesCompilerPass.php
@@ -11,11 +11,13 @@
 
 namespace Ynlo\GraphQLBundle\Component\TaggedServices;
 
+@trigger_error("TaggedServices component has been deprecated since v1.1 and will be deleted in v2.0, use symfony tag injection instead \"!tagged tag_name\"", E_USER_DEPRECATED);
+
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 /**
- * Class TaggedServicesCompilerPass
+ * @deprecated since v1.1 and will be deleted in v2.0, use symfony tag injection instead "!tagged tag_name"
  */
 class TaggedServicesCompilerPass implements CompilerPassInterface
 {