+++ CHANGELOG.md
@@ -50,6 +50,7 @@
  * Fixed duplicate field definition when interface publish a field with different name
  * Improve error handling, Add controlled errors, custom error formatter and handlers
  * Added support to display validation messages as errors
+ * Deprecated internal `TaggedServices` component in favor of symfony injection using "!tagged tag_name"
  
 1.0.6 - 18-03-23
 ----